refactor: move tabs and functions to separate folders
