+++ irw-dataset-builder/tabs/optional_data_tab.py
@@ -1,0 +1,19 @@
+import streamlit as st
+import pandas as pd
+
+def show_optional_data_tab():
+    st.markdown("""
+    ### Optional Data (Under Construction)
+
+    The **Optional Data** section is currently under construction.  
+    Possible mappings include:
+
+    - `rt` - Response time  
+    - `date` - Date/time of response  
+    - `qmatrix` - Q-matrix information  
+    - `rater` - Rater identification  
+    - `wave` - Wave/time point  
+    - `treat` - Treatment condition  
+
+    Stay tuned for updates as we continue to enhance this feature.
+    """)
