+++ irw-dataset-builder/utils/__init__.py
@@ -1,0 +1,8 @@
+from .data_processing import process_row_as_column, broadcasting_with_ranges
+from .highlighting import highlight_selection
+
+__all__ = [
+    'process_row_as_column',
+    'broadcasting_with_ranges',
+    'highlight_selection'
+]
